refactor: Round 2 — deep extraction across match_engine, world_ticker, and game_models
match_engine.py (1880 → 1679, -201):
  - Extract tag_match_simulator.py (189 lines) — tag match simulation loop
  - Extract match_db_integration.py — simulate_match_from_db + post-match angles
  - Extract match_rating.py (122 lines) — rating and heat calculations
  - MatchSimulator is now focused on singles/multi-person orchestration

world_ticker.py (1641 → 1231, -410):
  - Extract show_simulation_service.py (404 lines) — the 201-line _simulate_show
    God method plus promo evaluation and card psychology bonus
  - Extract tag_team_data.py (107 lines) — tag team name pools and generation
  - Extract ticker_query_helpers.py (35 lines) — DRY query helpers for
    active wrestlers, NPC feds, active feds

game_models.py (1372 → 173 barrel file, -1199):
  - Split into 5 focused model files:
    core_models.py (114) — WorldDB, PlayerDB, WorldStateDB
    federation_models.py (209) — GameFederationDB, PPVEventDB, BookingVisionDB
    wrestler_models.py (401) — GameWrestlerDB, WrestlerStatsDB, goals, history
    show_models.py (239) — ShowDB, MatchDB, ShowSegmentDB, PromoDB
    social_models.py (344) — StorylineDB, StableDB, ManagerDB, TagTeamDB
  - game_models.py is now a barrel re-export so all existing imports work

No behavior changes. All Python syntax validated, TypeScript compiles clean.
