+++ core_engine/match_engine.py
@@ -679,158 +679,9 @@ class MatchSimulator:
         )
 
     def _simulate_tag_match(self, participants: List[MatchParticipantState]) -> MatchResult:
-        """Simulate a tag team match with legal man tracking, tags, hot tags, and double-team spots."""
-        team_map: Dict[int, List[MatchParticipantState]] = {}
-        for p in participants:
-            team_id = p.team if p.team is not None else 0
-            team_map.setdefault(team_id, []).append(p)
-
-        team_ids = sorted(team_map.keys())
-        if len(team_ids) < 2:
-            return self._simulate_singles(participants)
-
-        # teams[0] = team A roster, teams[1] = team B roster
-        teams = [team_map[team_ids[0]], team_map[team_ids[1]]]
-
-        # Mutable state: legal_indices[i] = index of legal man in teams[i]
-        legal_indices = [0, 0]
-        # ticks_in[i] = how long team i's legal man has been in the ring
-        ticks_in = [0, 0]
-
-        # Tag matches run slightly longer
-        min_len, max_len = self.MATCH_LENGTH.get(self.card_position, (10, 20))
-        target_length = random.randint(min_len + 3, max_len + 5)
-
-        # Who's on offense: 0 = teams[0] attacks, 1 = teams[1] attacks
-        attacking_team = 0
-
-        while self.tick < target_length + 10:
-            self.tick += 1
-            ticks_in[0] += 1
-            ticks_in[1] += 1
-
-            atk_team = attacking_team
-            def_team = 1 - attacking_team
-            attacker = teams[atk_team][legal_indices[atk_team]]
-            defender = teams[def_team][legal_indices[def_team]]
-
-            # --- Tag-in, hot tag, and double-team opportunities ---
-            result = self._tag_team_action(
-                teams, legal_indices, ticks_in, attacking_team,
-                attacker, defender,
-            )
-            if result is not None:
-                action_type, attacker, defender, attacking_team = result
-                if action_type == "double_team":
-                    continue  # Double-team already applied; skip normal spot
-
-            # --- Normal spot ---
-            spot = self._generate_spot(attacker, defender)
-            self.spots.append(spot)
-            self._apply_spot(spot, attacker, defender)
-
-            if not attacker.finisher_available and attacker.momentum > FINISHER_MOMENTUM_THRESHOLD:
-                attacker.finisher_available = True
-
-            if self.tick >= target_length - 3:
-                finish_spot = self._attempt_finish(attacker, defender)
-                if finish_spot:
-                    self.spots.append(finish_spot)
-                    return self._build_result(finish_spot, attacker, defender, participants)
-
-            if self.tick > target_length * 0.6 and random.random() < NEAR_FALL_CHANCE:
-                near_fall = self._near_fall(attacker, defender)
-                if near_fall:
-                    self.spots.append(near_fall)
-
-            if self._should_switch_control(attacker, defender):
-                attacking_team = 1 - attacking_team
-
-        return MatchResult(
-            winner_id=None,
-            finish_type="time_limit_draw",
-            finish_description="The tag team match ends in a time limit draw!",
-            match_rating=self._calculate_rating(participants),
-            crowd_heat=self._calculate_heat(),
-            duration_ticks=self.tick,
-            spots=self.spots,
-        )
-
-    def _tag_team_action(
-        self,
-        teams: List[List[MatchParticipantState]],
-        legal_indices: List[int],
-        ticks_in: List[int],
-        attacking_team: int,
-        attacker: MatchParticipantState,
-        defender: MatchParticipantState,
-    ) -> Optional[tuple]:
-        """Handle tag-in, hot tag, and double-team logic for either team.
-
-        Returns ``None`` if no tag action occurred, or a tuple of
-        ``(action_type, attacker, defender, attacking_team)`` reflecting the
-        (possibly updated) match state.
-        """
-        atk_team = attacking_team
-        def_team = 1 - attacking_team
-
-        # --- Tag-in opportunity (attacking team tags to bring fresh partner) ---
-        if (ticks_in[atk_team] > TAG_IN_TICKS_THRESHOLD
-                and attacker.stamina < TAG_IN_STAMINA_THRESHOLD
-                and random.random() < TAG_IN_CHANCE
-                and len(teams[atk_team]) > 1):
-            legal_indices[atk_team] = (legal_indices[atk_team] + 1) % len(teams[atk_team])
-            ticks_in[atk_team] = 0
-            new_wrestler = teams[atk_team][legal_indices[atk_team]]
-            new_wrestler.momentum = min(100, new_wrestler.momentum + TAG_IN_MOMENTUM_BOOST)
-            self.spots.append(MatchSpot(
-                tick=self.tick, attacker_id=new_wrestler.wrestler_id,
-                defender_id=defender.wrestler_id, move_name="Tag",
-                move_type="tag", damage=0,
-                crowd_reaction="Tag made!", heat_change=1,
-                description=f"{attacker.name} {random.choice(TAG_DESCRIPTIONS)}! {new_wrestler.name} enters the ring!",
-            ))
-            return ("tag_in", new_wrestler, defender, attacking_team)
-
-        # --- Hot tag mechanic (defending team, beaten down, makes desperate tag) ---
-        if (ticks_in[def_team] > HOT_TAG_TICKS_THRESHOLD
-                and defender.health < HOT_TAG_HEALTH_THRESHOLD
-                and random.random() < HOT_TAG_CHANCE
-                and len(teams[def_team]) > 1):
-            legal_indices[def_team] = (legal_indices[def_team] + 1) % len(teams[def_team])
-            ticks_in[def_team] = 0
-            hot_tag = teams[def_team][legal_indices[def_team]]
-            hot_tag.momentum = min(100, hot_tag.momentum + HOT_TAG_MOMENTUM_BOOST)
-            hot_tag.finisher_available = True
-            self.spots.append(MatchSpot(
-                tick=self.tick, attacker_id=hot_tag.wrestler_id,
-                defender_id=attacker.wrestler_id, move_name="Hot Tag",
-                move_type="tag", damage=0,
-                crowd_reaction="The crowd erupts for the hot tag!",
-                heat_change=4,
-                description=f"{defender.name} desperately reaches out... HOT TAG! {hot_tag.name} storms into the ring on fire!",
-            ))
-            # Offense switches to the defending team
-            new_attacking_team = def_team
-            return ("hot_tag", hot_tag, attacker, new_attacking_team)
-
-        # --- Double-team opportunity (both partners briefly in ring) ---
-        if random.random() < DOUBLE_TEAM_CHANCE and len(teams[atk_team]) > 1:
-            partner = teams[atk_team][(legal_indices[atk_team] + 1) % len(teams[atk_team])]
-            move_name, dmg = random.choice(DOUBLE_TEAM_MOVES)
-            dt_spot = MatchSpot(
-                tick=self.tick, attacker_id=attacker.wrestler_id,
-                defender_id=defender.wrestler_id, move_name=move_name,
-                move_type="power", damage=dmg,
-                crowd_reaction="Incredible double-team!",
-                heat_change=3,
-                description=f"{attacker.name} and {partner.name} hit a {move_name} on {defender.name}!",
-            )
-            self.spots.append(dt_spot)
-            self._apply_spot(dt_spot, attacker, defender)
-            return ("double_team", attacker, defender, attacking_team)
-
-        return None
+        """Simulate a tag team match.  Delegates to tag_match_simulator module."""
+        from core_engine.tag_match_simulator import simulate_tag_match
+        return simulate_tag_match(self, participants)
 
     def _generate_spot(self, attacker: MatchParticipantState,
                        defender: MatchParticipantState) -> MatchSpot:
@@ -1368,70 +1219,18 @@ class MatchSimulator:
         return None
 
     def _calculate_rating(self, participants: List[MatchParticipantState]) -> float:
-        """Calculate match star rating (0.0 - 5.0)."""
-        # Base: average of participants' psychology and selling
-        avg_psychology = sum(p.stats.get("psychology", 50) for p in participants) / len(participants)
-        avg_selling = sum(p.stats.get("selling", 50) for p in participants) / len(participants)
-
-        base_quality = (avg_psychology + avg_selling) / 100  # 0-1 scale
-
-        # Bonus for spot variety
-        move_types_used = set(s.move_type for s in self.spots if not s.was_reversed)
-        variety_bonus = min(len(move_types_used) * VARIETY_BONUS_PER_TYPE, VARIETY_BONUS_CAP)
-
-        # Bonus for near falls
-        near_falls = sum(1 for s in self.spots if s.is_near_fall)
-        near_fall_bonus = min(near_falls * NEAR_FALL_BONUS_WEIGHT, NEAR_FALL_BONUS_CAP)
-
-        # Bonus for reversals (back-and-forth action)
-        reversals = sum(1 for s in self.spots if s.was_reversed)
-        reversal_bonus = min(reversals * REVERSAL_BONUS_WEIGHT, REVERSAL_BONUS_CAP)
-
-        # Match length bonus
-        length_bonus = min(self.tick / LENGTH_BONUS_DIVISOR, LENGTH_BONUS_CAP) if self.tick > LENGTH_BONUS_MIN_TICKS else 0
-
-        # Title match bonus
-        title_bonus = TITLE_MATCH_RATING_BONUS if self.is_title_match else 0
-
-        # Rivalry heat bonus — hot feuds produce better matches
-        rivalry_bonus = min(RIVALRY_HEAT_RATING_CAP, self.rivalry_heat / RIVALRY_HEAT_RATING_DIVISOR)
-
-        # Interference penalty — cheating cheapens ratings slightly
-        interference_penalty = INTERFERENCE_RATING_PENALTY if self._interference_happened else 0
-
-        # Signature move bonus — fans love seeing signature spots
-        sig_spots = sum(1 for s in self.spots if "Signature" in s.description or "signature" in s.description)
-        sig_bonus = min(sig_spots * SIGNATURE_BONUS_WEIGHT, SIGNATURE_BONUS_CAP)
-
-        # Stipulation bonus — gimmick matches get a bump for spectacle
-        stip_bonus = 0.0
-        stip = (self.stipulation or "").lower()
-        if "cage" in stip or "cell" in stip:
-            stip_bonus = CAGE_CELL_STIP_BONUS
-        elif "ladder" in stip or "table" in stip:
-            stip_bonus = LADDER_TABLE_STIP_BONUS
-        elif stip and stip not in ("", "standard"):
-            stip_bonus = GENERIC_STIP_BONUS
-
-        # Venue atmosphere bonus (set by caller via show_momentum scaling)
-        venue_tier = get_venue_tier(self.show_momentum * 100)  # approximate
-        venue_mod = VENUE_ATMOSPHERE.get(venue_tier, {}).get("rating_mod", 0)
-
-        rating = ((base_quality * 2.5) + variety_bonus + near_fall_bonus +
-                  reversal_bonus + length_bonus + title_bonus + rivalry_bonus +
-                  interference_penalty + sig_bonus + stip_bonus + venue_mod)
-        rating = min(RATING_MAX, max(RATING_MIN, rating + random.uniform(RATING_NOISE_MIN, RATING_NOISE_MAX)))
-        return round(rating, 1)
+        """Calculate match star rating (0.0 - 5.0).  Delegates to match_rating module."""
+        from core_engine.match_rating import calculate_rating
+        return calculate_rating(
+            self.spots, self.is_title_match, self.rivalry_heat,
+            self._interference_happened, self.stipulation,
+            self.show_momentum, self.tick, participants,
+        )
 
     def _calculate_heat(self) -> int:
-        """Calculate final crowd heat level."""
-        # Start from show momentum (crowd energy carried from prior segments)
-        base = self.show_momentum
-        # Rivalry heat gives a crowd bonus — fans care about these two
-        base += int(self.rivalry_heat * RIVALRY_HEAT_CROWD_FACTOR)
-        for spot in self.spots:
-            base += spot.heat_change
-        return max(CROWD_HEAT_MIN, min(CROWD_HEAT_MAX, base))
+        """Calculate final crowd heat level.  Delegates to match_rating module."""
+        from core_engine.match_rating import calculate_heat
+        return calculate_heat(self.spots, self.show_momentum, self.rivalry_heat)
 
     def _build_result(self, finish_spot: MatchSpot, attacker: MatchParticipantState,
                       defender: MatchParticipantState,
@@ -1504,377 +1303,12 @@ class MatchSimulator:
 
 
 # ---------------------------------------------------------------------------
-# Helper: Stat modifier calculation
+# DB integration functions — delegated to match_db_integration module.
+# Re-exported here for backward compatibility.
 # ---------------------------------------------------------------------------
 
-def _build_stat_modifiers(morale: float, ring_rust_days: int, conditioning: float) -> float:
-    """Compute a combined stat modifier from morale, ring rust, and conditioning.
-
-    Returns a multiplier (roughly 0.6-1.15) applied to all wrestler stats
-    before a match simulation.
-    """
-    # Morale affects performance: range MORALE_MODIFIER_BASE to BASE+RANGE
-    stat_modifier = MORALE_MODIFIER_BASE + (morale / 100) * MORALE_MODIFIER_RANGE
-
-    # Ring rust penalty for wrestlers who haven't competed recently
-    if ring_rust_days > RING_RUST_THRESHOLD_DAYS:
-        stat_modifier *= max(RING_RUST_FLOOR, 1.0 - (ring_rust_days / RING_RUST_DIVISOR))
-
-    # Conditioning modifier
-    cond_modifier = CONDITIONING_MODIFIER_BASE + (conditioning / 100) * CONDITIONING_MODIFIER_RANGE
-    stat_modifier *= cond_modifier
-
-    return stat_modifier
-
-
-# ---------------------------------------------------------------------------
-# Helper: Run a match from DB models
-# ---------------------------------------------------------------------------
-
-def simulate_match_from_db(db: Session, match: MatchDB, game_date: str = None) -> MatchResult:
-    """Load match data from DB, simulate, and persist results."""
-    participants_db = db.query(MatchParticipantDB).filter(
-        MatchParticipantDB.match_id == match.id,
-        MatchParticipantDB.role == "competitor",
-    ).all()
-
-    if len(participants_db) < 2:
-        return MatchResult(narrative_summary="Not enough competitors")
-
-    # Build participant states with morale modifier
-    participant_states = []
-    for p in participants_db:
-        wrestler = db.query(GameWrestlerDB).filter(GameWrestlerDB.id == p.wrestler_id).first()
-        stats = db.query(WrestlerStatsDB).filter(WrestlerStatsDB.wrestler_id == p.wrestler_id).first()
-
-        if not wrestler or not stats:
-            continue
-
-        morale = wrestler.morale if wrestler.morale is not None else 50
-        ring_rust = getattr(wrestler, "ring_rust_days", 0) or 0
-        conditioning = getattr(stats, "conditioning_level", 70) or 70
-        stat_modifier = _build_stat_modifiers(morale, ring_rust, conditioning)
-
-        # Load signature moves
-        sig_moves = []
-        if wrestler.signature_moves:
-            for sig in wrestler.signature_moves:
-                if isinstance(sig, (list, tuple)) and len(sig) >= 3:
-                    sig_moves.append(tuple(sig))
-
-        # Load charisma style
-        personality = wrestler.personality_traits or {}
-        charisma_style = personality.get("charisma_style", "humble") if isinstance(personality, dict) else "humble"
-
-        state = MatchParticipantState(
-            wrestler_id=wrestler.id,
-            name=wrestler.name,
-            health=wrestler.condition,  # Current condition affects match health
-            stats={
-                "power": int(stats.power * stat_modifier),
-                "technical": int(stats.technical * stat_modifier),
-                "aerial": int(stats.aerial * stat_modifier),
-                "brawling": int(stats.brawling * stat_modifier),
-                "submission": int(stats.submission * stat_modifier),
-                "stamina": int(stats.stamina * stat_modifier),
-                "toughness": int(stats.toughness * stat_modifier),
-                "speed": int(stats.speed * stat_modifier),
-                "charisma": int(stats.charisma * stat_modifier),
-                "psychology": int(stats.psychology * stat_modifier),
-                "selling": int(stats.selling * stat_modifier),
-            },
-            finisher_name=wrestler.finisher_name or "Finisher",
-            alignment=wrestler.alignment or "face",
-            team=p.team,
-            signature_moves=sig_moves,
-            charisma_style=charisma_style,
-        )
-        participant_states.append(state)
-
-    if len(participant_states) < 2:
-        return MatchResult(narrative_summary="Not enough competitors with stats")
-
-    # Determine card position (may be set by world ticker, fallback to heuristic)
-    card_position = getattr(match, "card_position", None) or "midcard"
-    if match.is_title_match and card_position == "midcard":
-        card_position = "main_event"
-
-    # Load manager context for wrestlers at ringside
-    managers = []
-    try:
-        from models.game_models import ManagerClientDB, ManagerDB
-        for p_state in participant_states:
-            bond = db.query(ManagerClientDB).filter_by(
-                client_wrestler_id=p_state.wrestler_id, is_active=True
-            ).first()
-            if bond:
-                mgr = db.query(ManagerDB).filter_by(id=bond.manager_id).first()
-                if mgr:
-                    managers.append(ManagerContext(
-                        manager_id=mgr.id,
-                        manager_name=mgr.name,
-                        client_wrestler_id=p_state.wrestler_id,
-                        interference_skill=mgr.interference_skill or 50,
-                        cunning=mgr.cunning or 50,
-                        specialization=bond.specialization or "all_around",
-                    ))
-    except Exception:
-        pass  # Manager integration is optional
-
-    # Calculate rivalry heat and trust between participants
-    rivalry_heat = 0
-    trust_penalty = 0.0
-    try:
-        from models.game_models import WrestlerRelationshipDB
-        if len(participant_states) >= 2 and match.world_id:
-            rel = db.query(WrestlerRelationshipDB).filter(
-                WrestlerRelationshipDB.world_id == match.world_id,
-                WrestlerRelationshipDB.wrestler1_id.in_(
-                    [participant_states[0].wrestler_id, participant_states[1].wrestler_id]
-                ),
-                WrestlerRelationshipDB.wrestler2_id.in_(
-                    [participant_states[0].wrestler_id, participant_states[1].wrestler_id]
-                ),
-            ).first()
-            if rel:
-                rivalry_heat = rel.rivalry_heat or 0
-                # Low trust = higher botch chance (wrestlers don't protect each other)
-                trust = rel.trust_level if rel.trust_level is not None else 50
-                if trust < LOW_TRUST_THRESHOLD:
-                    trust_penalty = (LOW_TRUST_THRESHOLD - trust) / LOW_TRUST_PENALTY_DIVISOR
-    except Exception:
-        pass  # Rivalry heat is optional
-
-    # Load frustration and morale for going-into-business checks
-    try:
-        from models.game_models import WrestlerGoalDB
-        for p_state in participant_states:
-            wrestler = db.query(GameWrestlerDB).filter(
-                GameWrestlerDB.id == p_state.wrestler_id
-            ).first()
-            if wrestler:
-                p_state._morale = wrestler.morale or 50
-                # Get max frustration from active goals
-                max_frust = db.query(WrestlerGoalDB).filter(
-                    WrestlerGoalDB.wrestler_id == p_state.wrestler_id,
-                    WrestlerGoalDB.status == "active",
-                ).all()
-                p_state._frustration = max((g.frustration for g in max_frust), default=0)
-                # Get ego from backstory
-                from models.game_models import WrestlerBackstoryDB
-                backstory = db.query(WrestlerBackstoryDB).filter(
-                    WrestlerBackstoryDB.wrestler_id == p_state.wrestler_id
-                ).first()
-                if backstory and backstory.real_personality:
-                    p_state.stats["ego"] = backstory.real_personality.get("ego", 50)
-    except Exception:
-        pass  # Frustration/ego loading is optional
-
-    # Show momentum passed via match attribute (set by world_ticker)
-    show_momentum = getattr(match, "_show_momentum", 50)
-
-    # Simulate
-    simulator = MatchSimulator(
-        planned_winner_id=match.winner_id,  # Pre-planned winner from booker
-        planned_finish=match.finish_type,
-        card_position=card_position,
-        is_title_match=match.is_title_match,
-        stipulation=match.stipulation,
-        managers=managers,
-        rivalry_heat=rivalry_heat,
-        show_momentum=show_momentum,
-    )
-    simulator._trust_penalty = trust_penalty
-    result = simulator.simulate(participant_states)
-
-    # Apply chemistry bonus from wrestler relationships
-    try:
-        from core_engine.match_aftermath import get_chemistry_bonus
-        if len(participant_states) >= 2 and match.world_id:
-            chem_bonus = get_chemistry_bonus(
-                db, match.world_id,
-                participant_states[0].wrestler_id,
-                participant_states[1].wrestler_id,
-            )
-            if chem_bonus > 0:
-                result.match_rating = round(
-                    min(5.0, result.match_rating + chem_bonus), 1
-                )
-    except Exception:
-        pass  # Chemistry bonus is optional enhancement
-
-    # Persist results back to DB
-    match.winner_id = result.winner_id
-    match.finish_type = result.finish_type
-    match.finish_description = result.finish_description
-    match.match_rating = result.match_rating
-    match.crowd_heat = result.crowd_heat
-    match.duration_minutes = result.duration_ticks  # 1 tick ≈ 1 minute
-    match.is_completed = True
-    match.simulation_log = [
-        {
-            "tick": s.tick, "attacker": s.attacker_id, "defender": s.defender_id,
-            "move": s.move_name, "move_type": s.move_type, "damage": s.damage,
-            "reversed": s.was_reversed, "is_near_fall": s.is_near_fall,
-            "is_finisher": s.is_finisher, "is_finish": s.is_finish,
-            "crowd_reaction": s.crowd_reaction,
-            "highlight_tier": (3 if s.is_finisher or s.is_finish else
-                               2 if s.is_near_fall or s.was_reversed or s.damage >= HIGHLIGHT_DAMAGE_THRESHOLD else 1),
-            "description": s.description,
-            "is_botch": s.is_botch,
-            "botch_severity": s.botch_severity,
-            "is_shoot": s.is_shoot,
-        }
-        for s in result.spots
-    ]
-
-    # Persist individual match events
-    for spot in result.spots:
-        db.add(MatchEventDB(
-            match_id=match.id,
-            tick=spot.tick,
-            acting_wrestler_id=spot.attacker_id,
-            target_wrestler_id=spot.defender_id,
-            event_type=spot.move_type,
-            description=spot.description,
-            crowd_reaction=spot.crowd_reaction,
-            heat_change=spot.heat_change,
-            damage=spot.damage,
-        ))
-
-    # Update winner's participation record
-    for p in participants_db:
-        if p.wrestler_id == result.winner_id:
-            p.is_winner = True
-        p.performance_rating = result.match_rating
-
-    # Wear on wrestlers' condition
-    for p_state in participant_states:
-        wrestler = db.query(GameWrestlerDB).filter(
-            GameWrestlerDB.id == p_state.wrestler_id
-        ).first()
-        if wrestler:
-            stamina_loss = int((100 - p_state.stamina) * STAMINA_WEAR_FACTOR)
-            health_loss = int((100 - p_state.health) * HEALTH_WEAR_FACTOR)
-            wrestler.condition = max(0, wrestler.condition - stamina_loss - health_loss)
-
-            # Injury risk check
-            if p_state.health < INJURY_HEALTH_THRESHOLD:
-                injury_prone = DEFAULT_INJURY_PRONE
-                s = db.query(WrestlerStatsDB).filter(
-                    WrestlerStatsDB.wrestler_id == wrestler.id
-                ).first()
-                if s:
-                    injury_prone = s.injury_prone
-                if random.random() < injury_prone / INJURY_PRONE_DIVISOR:
-                    wrestler.is_injured = True
-                    weeks = random.randint(2, 12)
-                    from game_service.world_ticker import advance_game_date
-                    # Use the match's actual game date for return date calculation
-                    match_date = game_date or getattr(match, "game_date", None)
-                    if not match_date and match.world_id:
-                        from models.game_models import WorldDB
-                        world = db.query(WorldDB).filter(WorldDB.id == match.world_id).first()
-                        if world:
-                            match_date = world.current_game_date
-                    if match_date:
-                        wrestler.injury_return_date = advance_game_date(
-                            match_date, weeks * 7
-                        )
-                    else:
-                        logger.warning("No game_date available for injury return date; skipping return date")
-                        wrestler.injury_return_date = None
-
-    # Generate post-match angle (faction run-ins, beatdowns)
-    result.post_match_angle = _generate_post_match_angle(
-        db, match, result, participant_states
-    )
-
-    return result
-
-
-def _generate_post_match_angle(
-    db: Session, match: MatchDB, result: MatchResult,
-    participants: List[MatchParticipantState],
-) -> Optional[Dict[str, Any]]:
-    """Check if a post-match angle occurs — faction attacks, saves, etc."""
-    if not result.winner_id or not match.world_id:
-        return None
-
-    try:
-        from models.game_models import StableMemberDB, StableDB, GameWrestlerDB
-
-        loser_id = None
-        for p in participants:
-            if p.wrestler_id != result.winner_id:
-                loser_id = p.wrestler_id
-                break
-        if not loser_id:
-            return None
-
-        # Check if winner is in a heel stable — faction beatdown chance
-        winner_member = db.query(StableMemberDB).filter_by(
-            wrestler_id=result.winner_id, is_active=True
-        ).first()
-        if winner_member:
-            stable = db.query(StableDB).filter_by(
-                id=winner_member.stable_id, is_active=True
-            ).first()
-            if stable and stable.alignment == "heel" and random.random() < FACTION_BEATDOWN_CHANCE:
-                # Faction beatdown on the loser
-                stablemates = db.query(StableMemberDB).filter_by(
-                    stable_id=stable.id, is_active=True
-                ).all()
-                attacker_ids = [m.wrestler_id for m in stablemates if m.wrestler_id != result.winner_id]
-                if attacker_ids:
-                    attackers = db.query(GameWrestlerDB).filter(
-                        GameWrestlerDB.id.in_(attacker_ids[:2])
-                    ).all()
-                    attacker_names = " & ".join(a.name for a in attackers)
-                    victim = db.query(GameWrestlerDB).filter_by(id=loser_id).first()
-                    desc = random.choice(POST_MATCH_ATTACK).format(
-                        attackers=f"{stable.name} ({attacker_names})",
-                        victim=victim.name if victim else "the loser"
-                    )
-                    return {
-                        "type": "faction_beatdown",
-                        "stable_id": stable.id,
-                        "stable_name": stable.name,
-                        "victim_id": loser_id,
-                        "attacker_ids": attacker_ids[:2],
-                        "description": desc,
-                    }
-
-        # Check if loser is in a face stable — save chance
-        loser_member = db.query(StableMemberDB).filter_by(
-            wrestler_id=loser_id, is_active=True
-        ).first()
-        if loser_member:
-            stable = db.query(StableDB).filter_by(
-                id=loser_member.stable_id, is_active=True
-            ).first()
-            if stable and stable.alignment == "face" and random.random() < FACTION_SAVE_CHANCE:
-                stablemates = db.query(StableMemberDB).filter_by(
-                    stable_id=stable.id, is_active=True
-                ).all()
-                saver_ids = [m.wrestler_id for m in stablemates if m.wrestler_id != loser_id]
-                if saver_ids:
-                    savers = db.query(GameWrestlerDB).filter(
-                        GameWrestlerDB.id.in_(saver_ids[:2])
-                    ).all()
-                    saver_names = " & ".join(s.name for s in savers)
-                    desc = random.choice(POST_MATCH_SAVE).format(
-                        savers=f"{stable.name} ({saver_names})"
-                    )
-                    return {
-                        "type": "faction_save",
-                        "stable_id": stable.id,
-                        "stable_name": stable.name,
-                        "saved_id": loser_id,
-                        "saver_ids": saver_ids[:2],
-                        "description": desc,
-                    }
-    except Exception as e:
-        logger.warning("Post-match angle generation failed: %s", e)
-
-    return None
+from core_engine.match_db_integration import (  # noqa: E402
+    _build_stat_modifiers,
+    simulate_match_from_db,
+    _generate_post_match_angle,
+)
